Refactor `Application` with PImpl and improve UI in `Primary`.
The `Application` class has been refactored to use the PImpl idiom for better encapsulation and reduced compilation dependencies. A new debug UI has been added to the `Primary` scene, featuring FPS tracking and elapsed time display. Minor code improvements include parameter type adjustments and cleanup for consistency.

+++ source/core/SceneManager.cpp
@@ -37,7 +37,7 @@ void SceneManager::popScene() {
   }
 }
 
-void SceneManager::update(const float dt) {
+void SceneManager::update(const double dt) {
   if (m_scenes.empty()) {
     LOG_WARN("No scenes to update");
     return;
